requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt src/requirements/production.in src/requirements/test.in
#
apipkg==1.4               # via execnet
argh==0.26.2              # via watchdog
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
colorama==0.3.9           # via pytest-watch
coverage==4.4.1           # via pytest-cov, pytest-testmon
django-storages==1.6.5
django==1.11.4
djangorestframework==3.6.3
docopt==0.6.2             # via pytest-watch
execnet==1.4.1            # via pytest-xdist
idna==2.6                 # via requests
markdown==2.6.8
olefile==0.44             # via pillow
pathtools==0.1.2          # via watchdog
pillow==4.2.1
psycopg2==2.7
py==1.4.34                # via pytest, pytest-xdist
pytest-django==3.1.2
pytest-sugar==0.8.0
pytest-testmon==0.9.4
pytest-watch==4.1.0
pytest-xdist==1.16.0
pytest==3.0.7
pytest_cov==2.4.0
python-memcached==1.58
pytz==2017.2
pyyaml==3.12              # via watchdog
requests-mock==1.3.0
requests==2.18.4
six==1.10.0               # via python-memcached, requests-mock
termcolor==1.1.0          # via pytest-sugar
urllib3==1.22             # via requests
watchdog==0.8.3           # via pytest-watch
